better error message

+++ setup.sh
@@ -3,7 +3,11 @@
 set -eou pipefail
 
 # install certs
-mkcert -install -pkcs12 localhost
-cp localhost.p12 ./api/src/main/resources/
-cp localhost.p12 ./site/src/main/resources/
-rm localhost.p12
+if command -v mkcert &> /dev/null; then
+  mkcert -install -pkcs12 localhost
+  cp localhost.p12 ./api/src/main/resources/
+  cp localhost.p12 ./site/src/main/resources/
+  rm localhost.p12
+else
+  echo You must install mkcert
+fi
